Squashed commit of the following:
commit c64927e6def1ed31981a2f02d114de2dc02d7477
Author: Thaumy <thaumy@outlook.com>
Date:   Wed Aug 10 14:00:53 2022 +0800

    add log helper

commit 58d82549c9431d67a2a8698983c62da5f665952f
Author: Thaumy <thaumy@outlook.com>
Date:   Sat Aug 6 21:42:12 2022 +0800

    util.di

commit f9843c79631f2b1fc8c5bc1115969953c355003d
Author: Thaumy <thaumy@outlook.com>
Date:   Sat Aug 6 09:29:19 2022 +0800

    update dependency

commit 32fea89d98c0ff51ca3da1deced6811f4e93344f
Author: Thaumy <thaumy@outlook.com>
Date:   Sat Aug 6 02:59:04 2022 +0800

    wrapped types, getFstRow

commit f125f811b877fabcdb9df32acbbc20a098ea9285
Author: Thaumy <thaumy@outlook.com>
Date:   Fri Aug 5 18:21:02 2022 +0800

    pipeline interfaces

commit ae70714399b53361f079ec5ab4b9e971e3b0d621
Author: Thaumy <thaumy@outlook.com>
Date:   Fri Aug 5 11:27:15 2022 +0800

    update readme.md

commit f0029f6fee947136e9b634a5b4f58c3eec069691
Author: Thaumy <thaumy@outlook.com>
Date:   Thu Aug 4 20:54:08 2022 +0800

    update pipelines

commit 97424f2e581100515ff4b0f7eec8128d7fcf267b
Author: Thaumy <thaumy@outlook.com>
Date:   Wed Aug 3 10:55:28 2022 +0800

    updated pipelines

+++ pilipala-frame/container/comment.fs
@@ -16,7 +16,7 @@ type CommentData =
       CreateTime: DateTime
       Binding: CommentBinding
       UserId: u64
-      Permission: u16
+      Permission: u8
       Item: string -> Option'<obj> }
 
 /// 评论映射
@@ -27,7 +27,7 @@ type IMappedComment =
     abstract CreateTime: DateTime with get, set
     abstract Binding: CommentBinding with get, set
     abstract UserId: u64 with get, set
-    abstract Permission: u16 with get, set
+    abstract Permission: u8 with get, set
     abstract Item: string -> Option'<obj> with get, set
 //TODO 可以添加Item迭代器，以实现在Init管道的udf初始化
