+++ pilipala-frame/pipeline/comment/modify.fs
@@ -12,7 +12,7 @@ type ICommentModifyPipelineBuilder =
     abstract Binding: BuilderItem<u64 * CommentBinding>
     abstract CreateTime: BuilderItem<u64 * DateTime>
     abstract UserId: BuilderItem<u64 * u64>
-    abstract Permission: BuilderItem<u64 * u16>
+    abstract Permission: BuilderItem<u64 * u8>
     abstract Item: string -> BuilderItem<u64 * obj>
 
     //用于遍历Item
@@ -23,5 +23,5 @@ type ICommentModifyPipeline =
     abstract Binding: (u64 * CommentBinding) -> u64 * CommentBinding
     abstract CreateTime: (u64 * DateTime) -> u64 * DateTime
     abstract UserId: (u64 * u64) -> u64 * u64
-    abstract Permission: (u64 * u16) -> u64 * u16
+    abstract Permission: (u64 * u8) -> u64 * u8
     abstract Item: string -> Option'<u64 * obj -> u64 * obj>
